+++ test/dummy/config/importmap.rb
@@ -1,3 +1,4 @@
 pin "application"
+pin "clipboard_controller"
 pin "@hotwired/stimulus", to: "stimulus.min.js"
 pin "@hotwired/turbo-rails", to: "turbo.min.js"